Extract src-tauri/src/git.rs — everything Episko asks git
The largest tested block, and the easiest cut of the phase: the whole cluster is
contiguous in lib.rs, create_worktree through git_action, 1,006 lines. 15 of the
69 tests came with it. lib.rs 4,532 -> 3,026.

Stayed private to the module: same_path, git_cmd, git_run, upstream_state,
remove_worktree_impl, and all seven serde structs. Only git_repo_info crosses
back, because list_external_sessions calls it.

Also adds `#[cfg(test)] mod testutil;`, holding scratch_dir and its COUNTER.
Three future modules' tests need a fresh temp dir (git, transcript, usage), so
it can't live in any one of them and copying it is the drift bug PLAN records
for SORT_META. Only that: wt_root, git() and commit() have a single owner and
went into git.rs's own test mod.

Two mechanics confirmed by compiling rather than by reasoning, both open
questions in the kickoff prompt:

- pub(crate) works on a #[tauri::command] fn in a private module, so the
  "pub(crate), never pub" rule needs no exception. tasks.rs's plain `pub` reads
  like a counter-example but isn't — `mod tasks;` is private, so pub inside it
  is unreachable from outside the crate either way.
- Struct fields need no annotation at all. A private field is visible to the
  defining module and every descendant, and every module here is a descendant
  of the crate root, so git.rs reads state.sessions and s.workdir untouched.
  AppState/Session took pub(crate) only to satisfy private-in-public on the
  commands that name them. That is a second argument for keeping them in the
  crate root: a state.rs would have needed pub(crate) on every field.

Verification, both halves:
- git.rs vs the pre-slice lib.rs, line-sorted: identical, 1,377 non-blank lines,
  no differences at all.
- testutil.rs likewise identical (10 lines) once the 4-space dedent is
  normalised — it moved out of an inner `mod tests` to a top-level module.
- the remainder of lib.rs vs pre-slice-minus-those-ranges: 37 lines, each
  intended — mod git; / mod testutil; / use crate::git::git_repo_info;,
  pub(crate) on AppState and Session, the 12 git:: qualifications in
  generate_handler!, and the test mod swapping Path/PathBuf/Command/AtomicU32
  (used only by the moved git tests) for crate::testutil::{scratch_dir, COUNTER}.

The remainder check now uses diff(1). The hand-rolled resync it replaced
mis-aligned two insertions two lines apart and reported a desync that was not
one — the same class of false report PLAN already records twice.

cargo test: 69 passed, same as baseline. cargo check --tests clean, no
warnings. No TypeScript touched.
